[+] Allow one script updating

+++ slave/setup.sh
@@ -7,11 +7,24 @@ fi
 
 BIN_PATH='/usr/local/bin/qcheckserv'
 
-echo "@INFO: Installing qcheckserv"
 if [ ! -d "${BIN_PATH}"  ]; then
-    mkdir "${BIN_PATH}"
+    echo "@INFO: Installing qcheckserv"
+else
+    echo "@INFO: Updating qcheckserv"
+    rm -rf "${BIN_PATH}/"
 fi
-cp ./.env_example "${BIN_PATH}/.env"
+mkdir "${BIN_PATH}"
+
+if [ -f "${BIN_PATH}/.env" ]; then
+    read -p "> .env file detected. Override with fresh? [y/N]" yn
+    case $yn in
+        [Yy]* ) echo "@INFO: Replacing config"; cp ./.env_example "${BIN_PATH}/.env"; break;;
+        * ) echo "@INFO: Leaving old config";;
+    esac
+else
+    cp ./.env_example "${BIN_PATH}/.env"
+fi
+
 cp ./server_checker.py "${BIN_PATH}/"
 
 echo "@INFO: Copying service files"
@@ -20,5 +33,6 @@ cp ./services/qcheckserv_server_checker.timer /etc/systemd/system/
 
 echo "@INFO: Reloading services"
 systemctl daemon-reload
+systemctl enable qcheckserv_server_checker.timer
 systemctl start qcheckserv_server_checker.timer
 systemctl status qcheckserv_server_checker.timer
